classifier.py
"""
Simple Image Classifier for Injury Detection
Python 3.14 Compatible - No NumPy Required
Uses pure Python for all calculations
"""

from PIL import Image
import io

class InjuryClassifier:
    """
    Simple demo classifier that analyzes image features
    In production, replace with a trained CNN model
    
    This version uses pure Python (no NumPy) for Python 3.14 compatibility
    """
    
    def __init__(self):
        self.categories = ['minor_cut', 'burn', 'abrasion', 'bruise', 'swelling', 'unknown']
        
    def preprocess_image(self, image_bytes):
        """Convert image bytes to processable format"""
        try:
            image = Image.open(io.BytesIO(image_bytes))
            # Convert to RGB if necessary
            if image.mode != 'RGB':
                image = image.convert('RGB')
            # Resize to standard size
            image = image.resize((224, 224))
            return image
        except Exception as e:
            print(f"Error preprocessing image: {e}")
            return None
    
    def analyze_color_features(self, image):
        """
        Analyze color characteristics using pure Python
        No NumPy dependency required
        """
        # Get all pixels as list of (R, G, B) tuples
        pixels = list(image.getdata())
        num_pixels = len(pixels)
        
        if num_pixels == 0:
            return {
                'avg_red': 0,
                'avg_green': 0,
                'avg_blue': 0,
                'red_var': 0,
                'red_dominance': 0
            }
        
        # Calculate average color values
        total_r = sum(p[0] for p in pixels)
        total_g = sum(p[1] for p in pixels)
        total_b = sum(p[2] for p in pixels)
        
        avg_red = total_r / num_pixels
        avg_green = total_g / num_pixels
        avg_blue = total_b / num_pixels
        
        # Calculate variance for red channel
        mean_r = avg_red
        red_variance = sum((p[0] - mean_r) ** 2 for p in pixels) / num_pixels
        
        # Calculate red dominance (how much redder than other channels)
        red_dominance = avg_red - (avg_green + avg_blue) / 2
        
        return {
            'avg_red': avg_red,
            'avg_green': avg_green,
            'avg_blue': avg_blue,
            'red_var': red_variance,
            'red_dominance': red_dominance
        }
    
    def classify(self, image_bytes):
        """
        Classify injury from image bytes
        Returns: (category, confidence, features)
        
        NOTE: This is a DEMO implementation using simple heuristics.
        In production, replace with a trained deep learning model.
        """
        image = self.preprocess_image(image_bytes)
        
        if image is None:
            return 'unknown', 0.5, {}
        
        features = self.analyze_color_features(image)
        
        # Simple heuristic classification (DEMO ONLY)
        # In production, use a trained CNN model here
        
        red_dominance = features['red_dominance']
        red_var = features['red_var']
        avg_red = features['avg_red']
        
        # These are placeholder heuristics for demo purposes
        if red_dominance > 30 and red_var > 1000:
            # High red variance might indicate burn
            category = 'burn'
            confidence = 0.65
        elif red_dominance > 20 and avg_red > 140:
            # Red with some variance might be a cut
            category = 'minor_cut'
            confidence = 0.60
        elif red_dominance > 10:
            # Moderate redness might be abrasion
            category = 'abrasion'
            confidence = 0.58
        elif features['avg_blue'] > features['avg_red'] and features['avg_red'] < 100:
            # Bluish tint might indicate bruising
            category = 'bruise'
            confidence = 0.62
        else:
            # Default to unknown for safety
            category = 'unknown'
            confidence = 0.50
        
        return category, confidence, features
    
    def get_training_recommendation(self):
        """
        Returns information about training a real model
        """
        return """
        TO TRAIN A PRODUCTION MODEL FOR PYTHON 3.14:
        
        1. Collect a dataset of labeled injury images:
           - Minor cuts (500+ images)
           - Burns (500+ images)
           - Abrasions (500+ images)
           - Bruises (500+ images)
           - Swelling (500+ images)
        
        2. Use transfer learning with pre-trained models:
           - MobileNetV2 (lightweight, good for web)
           - ResNet50 (more accurate, heavier)
           - EfficientNet (best balance)
        
        3. Python 3.14 Compatible Frameworks:
           - TensorFlow 2.15+ 
           - PyTorch 2.1+
           - ONNX Runtime (for inference)
        
        4. Training code example (TensorFlow):
           ```python
           from tensorflow.keras.applications import MobileNetV2
           from tensorflow.keras.layers import Dense, GlobalAveragePooling2D
           from tensorflow.keras.models import Model
           
           base_model = MobileNetV2(weights='imagenet', include_top=False)
           x = base_model.output
           x = GlobalAveragePooling2D()(x)
           x = Dense(128, activation='relu')(x)
           predictions = Dense(5, activation='softmax')(x)
           
           model = Model(inputs=base_model.input, outputs=predictions)
           ```
        
        5. Deploy the trained model:
           - Save as TensorFlow SavedModel or ONNX
           - Load in this classifier's __init__ method
           - Replace classify() method with model.predict()
        
        IMPORTANT: For medical applications, ensure proper validation,
        testing, and compliance with healthcare regulations.
        
        NOTE: This demo version uses pure Python (no NumPy) for
        maximum compatibility with Python 3.14.
        """


# For demo purposes, we'll create a simple rule-based classifier
def create_classifier():
    """Factory function to create classifier instance"""
    return InjuryClassifier()
